Fixed a bug in the initialization of the CC2531 dongle
The CC2531 dongle needs to be reinitialized first by sending a CMD_INIT message. This
was missing from the previous versions but the firmware was quite working even without
this message. Anyway, this message is now sent each time the device is used, and this
fixes some issues such as listening on a wrong channel.